Restructure HashStore module, 'HashStore' is now an interface, 'FileHashStore' implements 'HashStore', add new class 'HashStoreFactory' and update pytests
