+++ src/hashstore/hashstore_config.py
@@ -2,9 +2,10 @@
 # Default path for `FileHashStore` if no path is provided
 STORE_PATH = "/var/filehashstore/"
 # Desired amount of directories when sharding an object to form the permanent address
-DIR_DEPTH = 3 # DO NOT CHANGE UNLESS SETTING UP NEW HASHSTORE
+DIR_DEPTH = 3 # WARNING: DO NOT CHANGE UNLESS SETTING UP NEW HASHSTORE
 # Width of directories created when sharding an object to form the permanent address
-DIR_WIDTH = 2 # DO NOT CHANGE UNLESS SETTING UP NEW HASHSTORE
+DIR_WIDTH = 2 # WARNING: DO NOT CHANGE UNLESS SETTING UP NEW HASHSTORE
+# Format of the metadata
 SYSMETA_NS = "http://ns.dataone.org/service/types/v2.0"
 # Algorithm values supported by python hashlib 3.9.0+
 DEFAULT_ALGO_LIST = ["sha1", "sha256", "sha384", "sha512", "md5"]
