+++ tests/test_hashstore.py
@@ -1,14 +1,14 @@
 """Test module for HashStore (and HashStoreFactory)"""
 import importlib.metadata
 import pytest
-from hashstore import HashStore
-from hashstore.hashstore import HashStoreFactory, FileHashStore
+from hashstore.filehashstore.filehashstore import FileHashStore
+from hashstore.hashstore_factory import HashStoreFactory
 
 
 @pytest.fixture(name="store")
 def init_store():
     """Create store path for all tests"""
-    store = HashStore()
+    store = FileHashStore()
     return store
 
 
@@ -23,7 +23,7 @@ def test_factory(tmp_path):
     directory = tmp_path / "metacat"
     directory.mkdir()
     # Get factory and hashstore
-    factory = HashStore.hashstore_factory
+    factory = HashStoreFactory()
     assert isinstance(factory, HashStoreFactory)
 
 
@@ -31,11 +31,10 @@ def test_factory_get_hashstore_filehashstore(tmp_path):
     """Check factory creates instance of FileHashStore"""
     directory = tmp_path / "metacat"
     directory.mkdir()
-    hashstore_path = directory.as_posix()
     hashstore_type = "filehashstore"
     # Get factory and hashstore
-    factory = HashStore.hashstore_factory
-    store = factory.get_hashstore(hashstore_path, hashstore_type)
+    factory = HashStoreFactory()
+    store = factory.get_hashstore(hashstore_type)
     assert isinstance(store, FileHashStore)
 
 
@@ -43,8 +42,7 @@ def test_factory_get_hashstore_unsupported_fs(tmp_path):
     """Check that ValueError is raised when provided with unsupported store type"""
     directory = tmp_path / "metacat"
     directory.mkdir()
-    hashstore_path = directory.as_posix()
     hashstore_type = "s3hashstore"
-    factory = HashStore.hashstore_factory
+    factory = HashStoreFactory()
     with pytest.raises(ValueError):
-        factory.get_hashstore(hashstore_path, hashstore_type)
+        factory.get_hashstore(hashstore_type)
